# example/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/{tenantID}/projections": {
            "get": {
                "description": "Get the states of projections for a specific tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projection"
                ],
                "summary": "Get projection states",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "item"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Projection IDs",
                        "name": "projections",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/projection.StateResponseDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{tenantID}/projections/rebuild": {
            "post": {
                "description": "Rebuild specified projections for a tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projection"
                ],
                "summary": "Rebuild projections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "item"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Projections",
                        "name": "projections",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{tenantID}/projections/start": {
            "post": {
                "description": "Start specified projections for a tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projection"
                ],
                "summary": "Start projections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "item"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Projections",
                        "name": "projections",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{tenantID}/projections/stop": {
            "post": {
                "description": "Stop specified projections for a tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projection"
                ],
                "summary": "Stop projections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "item"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Projections",
                        "name": "projections",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{tenantID}/projections/{projectionID}": {
            "delete": {
                "description": "Delete a projection by its ID for a specific tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projection"
                ],
                "summary": "Delete a projection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"item\"",
                        "description": "Projection ID",
                        "name": "projectionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{tenantID}/seeders/items": {
            "post": {
                "description": "Seed a specific number of items for a tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seeder"
                ],
                "summary": "Seed items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1000,
                        "description": "Number of items to seed",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "projection.StateResponseDTO": {
            "type": "object",
            "properties": {
                "executionTimeOut": {
                    "type": "string"
                },
                "finishingTimeOut": {
                    "type": "string"
                },
                "hPatchStrategy": {
                    "type": "string"
                },
                "inputQueueLength": {
                    "type": "integer"
                },
                "preparationTimeOut": {
                    "type": "string"
                },
                "projectionID": {
                    "type": "string"
                },
                "projectionType": {
                    "type": "string"
                },
                "rebuildExecutionTimeOut": {
                    "type": "string"
                },
                "retryDurations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "state": {
                    "type": "string"
                },
                "tenantID": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}